+++ internal/eqemuserver/client.go
@@ -57,6 +57,48 @@ func (c *Client) GetZoneList() (WorldZoneList, error) {
 	return zoneList, nil
 }
 
+type LockStatusResponse struct {
+	Data struct {
+		Locked bool `json:"locked"`
+	} `json:"data"`
+	ExecutionTime string `json:"execution_time"`
+	Method        string `json:"method"`
+}
+
+// GetLockStatus returns the lock status of the server
+func (c *Client) GetLockStatus() (bool, error) {
+	o, err := c.telnet.Command(
+		telnet.CommandConfig{Command: "api lock_status", EnforceJson: true},
+	)
+	if err != nil {
+		return false, err
+	}
+
+	var r LockStatusResponse
+	err = json.Unmarshal([]byte(o), &r)
+	if err != nil {
+		return false, err
+	}
+
+	return r.Data.Locked, nil
+}
+
+// SetLockStatus returns the lock status of the server
+func (c *Client) SetLockStatus(locked bool) error {
+	command := "lock"
+	if !locked {
+		command = "unlock"
+	}
+	_, err := c.telnet.Command(
+		telnet.CommandConfig{Command: command},
+	)
+	if err != nil {
+		return err
+	}
+
+	return nil
+}
+
 type WorldClientList struct {
 	Data []struct {
 		AccountID            int    `json:"account_id"`